+++ demo/app.component.ts
@@ -56,7 +56,8 @@ import { Component } from '@angular/core';
             <ul>
               <li><a href="#" (click)="state='cell-selection'">Cell</a></li>
               <li><a href="#" (click)="state='single-selection'">Single Row</a></li>
-              <li><a href="#" (click)="state='multi-selection'">Mulit Row</a></li>
+              <li><a href="#" (click)="state='multi-selection'">Multi Row</a></li>
+              <li><a href="#" (click)="state='multi-click-selection'">Multi Click Row</a></li>
               <li><a href="#" (click)="state='multidisable-selection'">Disable Callback</a></li>
               <li><a href="#" (click)="state='chkbox-selection'">Checkbox</a></li>
             </ul>
@@ -111,6 +112,7 @@ import { Component } from '@angular/core';
         <multi-selection-demo *ngIf="state === 'multi-selection'"></multi-selection-demo>
         <multidisable-selection-demo *ngIf="state === 'multidisable-selection'"></multidisable-selection-demo>
         <chkbox-selection-demo *ngIf="state === 'chkbox-selection'"></chkbox-selection-demo>
+        <multi-click-selection-demo *ngIf="state === 'multi-click-selection'"></multi-click-selection-demo>
 
         <!-- Templates -->
         <template-ref-demo *ngIf="state === 'templateref'"></template-ref-demo>