+++ src/types/selection.type.ts
@@ -1,6 +1,7 @@
 export enum SelectionType {
   single = 'single' as any,
   multi = 'multi' as any,
+  multiClick = 'multiClick' as any,
   cell = 'cell' as any,
   checkbox = 'checkbox' as any
 }